+++ cmd/helm/testdata/testcharts/chart-for-template-output/templates/serviceaccount.yaml
@@ -1,0 +1,8 @@
+{{- if .Values.serviceAccount.create -}}
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: {{ include "chart-for-template-output.serviceAccountName" . }}
+  labels:
+{{ include "chart-for-template-output.labels" . | nindent 4 }}
+{{- end -}}
