+++ cmd/helm/testdata/testcharts/chart-for-template-output/templates/tests/test-connection.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: "{{ include "chart-for-template-output.fullname" . }}-test-connection"
+  labels:
+{{ include "chart-for-template-output.labels" . | nindent 4 }}
+  annotations:
+    "helm.sh/hook": test-success
+spec:
+  containers:
+    - name: wget
+      image: busybox
+      command: ['wget']
+      args:  ['{{ include "chart-for-template-output.fullname" . }}:{{ .Values.service.port }}']
+  restartPolicy: Never
